flake: switch the Discourse domain name in dev module
The previous domain name - "localhost" combined with some Nginx rules
caused frontend errors when logging in to Discourse web UI.

Replacing it with "localhost:8080" fixed problems with logging in, but
caused other problems with Discourse web UI - SVG icons were not
loading. Fortunately, this was the only problem was easily fixable by
adding a Nginx rewrite rule.

Whole setting of Discourse domain name in dev module is a bit hacky, but
it is only used in development environment, not in production (or future
tests).