packages: introduce verify-login
The package is a Python script that uses selenium to verify that the
provided credentials can be used to log into the website.

Assuming that the website is available at http://localhost:8080 (e.g. by
running the dev VM - `nix run .#vm`), the package can be used by the
developer with the following command:
    $ nix run .#verify-login -- http://localhost:8080