site settings: create students group
Usually Discourse entities such as groups are created by the site admin,
using the web interface. In my opinion, that causes issues in terms of
transparency, reproducibility and testing - looking at the code is not
enough to know what groups will be created, so it's not possible to test
the impact of specific group settings on the site before applying them.

This commit tries to address those issues by making configuration of
those entities declarative. It works by:
1. Extending the nixpkgs's Discourse module with a set of new options
   (`services.discourse.groups.*`). See `modules/discourse-extended.nix`
   for details.
2. Adding a new package, `applier`, that is responsible for applying
   options defined in `services.discourse.groups.*` to the site using
   Discourse's API. See `packages/applier` for details.