packages: generate a Discourse API python client
This is a failed attempt of OpenAPI client generation for the Discourse
API. It consists of:
- Generating the OpenAPI specification from the Discourse source code
- Generating a Python client from the OpenAPI specification using the
  swagger-codegen3 tool
- Packaging the generated client with Nix

It's failed because the generated client doesn't work. It throws
exceptions when trying to parse the JSON responses from the API. Upon
further investigation, it seems that either the generated OpenAPI
specification is invalid, or the swagger-codegen tool is buggy. I
suspect the latter, because after manually inspecting the generated
OpenAPI spec, the JSON responses from the API seem to adhere to it.

I also tried to generate a client using the `openapi-generator-cli`
(https://github.com/OpenAPITools/openapi-generator), but it failed to
produce any code.